refactor: change by codestyle
