+++ src/features/savanna/Savanna.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Savanna = () => {
+  return (
+    <div>
+      <div>Savanna</div>
+    </div>
+  );
+};
+
+export default Savanna;
